+++ tencentcloud/postgres/v20170312/client.go
@@ -633,12 +633,37 @@ func NewDestroyDBInstanceResponse() (response *DestroyDBInstanceResponse) {
     return
 }
 
-// 本接口 (DestroyDBInstance) 用于销毁指定DBInstanceId对应的实例。当前仅适用于按量计费实例。
+// 本接口 (DestroyDBInstance) 用于彻底下线指定DBInstanceId对应的实例，下线后实例数据将彻底删除，无法找回，只能下线隔离中的实例。
 func (c *Client) DestroyDBInstance(request *DestroyDBInstanceRequest) (response *DestroyDBInstanceResponse, err error) {
     if request == nil {
         request = NewDestroyDBInstanceRequest()
     }
     response = NewDestroyDBInstanceResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewDisIsolateDBInstancesRequest() (request *DisIsolateDBInstancesRequest) {
+    request = &DisIsolateDBInstancesRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("postgres", APIVersion, "DisIsolateDBInstances")
+    return
+}
+
+func NewDisIsolateDBInstancesResponse() (response *DisIsolateDBInstancesResponse) {
+    response = &DisIsolateDBInstancesResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 本接口（DisIsolateDBInstances）用于解隔离实例
+func (c *Client) DisIsolateDBInstances(request *DisIsolateDBInstancesRequest) (response *DisIsolateDBInstancesResponse, err error) {
+    if request == nil {
+        request = NewDisIsolateDBInstancesRequest()
+    }
+    response = NewDisIsolateDBInstancesResponse()
     err = c.Send(request, response)
     return
 }
@@ -739,6 +764,31 @@ func (c *Client) InquiryPriceUpgradeDBInstance(request *InquiryPriceUpgradeDBIns
         request = NewInquiryPriceUpgradeDBInstanceRequest()
     }
     response = NewInquiryPriceUpgradeDBInstanceResponse()
+    err = c.Send(request, response)
+    return
+}
+
+func NewIsolateDBInstancesRequest() (request *IsolateDBInstancesRequest) {
+    request = &IsolateDBInstancesRequest{
+        BaseRequest: &tchttp.BaseRequest{},
+    }
+    request.Init().WithApiInfo("postgres", APIVersion, "IsolateDBInstances")
+    return
+}
+
+func NewIsolateDBInstancesResponse() (response *IsolateDBInstancesResponse) {
+    response = &IsolateDBInstancesResponse{
+        BaseResponse: &tchttp.BaseResponse{},
+    }
+    return
+}
+
+// 本接口（IsolateDBInstances）用于隔离实例
+func (c *Client) IsolateDBInstances(request *IsolateDBInstancesRequest) (response *IsolateDBInstancesResponse, err error) {
+    if request == nil {
+        request = NewIsolateDBInstancesRequest()
+    }
+    response = NewIsolateDBInstancesResponse()
     err = c.Send(request, response)
     return
 }