+++ tencentcloud/trtc/v20190722/models.go
@@ -1130,7 +1130,7 @@ type OutputParams struct {
 	// 自定义录制文件名称前缀。请先在实时音视频控制台开通录制功能，https://cloud.tencent.com/document/product/647/50768
 	RecordId *string `json:"RecordId,omitempty" name:"RecordId"`
 
-	// 取值范围[0,1]，填0无实际含义; 填1：指定录制文件格式为mp3
+	// 取值范围[0,1]，填0无实际含义; 填1：指定录制文件格式为mp3。此参数不建议使用，建议在实时音视频控制台配置纯音频录制模板。
 	RecordAudioOnly *uint64 `json:"RecordAudioOnly,omitempty" name:"RecordAudioOnly"`
 }